Struct stderrlog::StdErrLog

pub struct StdErrLog { /* fields omitted */ }

Data specific to this logger

Implementations

impl StdErrLog

pub fn new() -> StdErrLog

creates a new stderr logger

pub fn verbosity(&mut self, verbosity: usize) -> &mut StdErrLog

Sets the verbosity level of messages that will be displayed

pub fn quiet(&mut self, quiet: bool) -> &mut StdErrLog

silence all output, no matter the value of verbosity

pub fn show_level(&mut self, levels: bool) -> &mut StdErrLog

Enables or disables the use of timestamps in log messages (default is true)

pub fn timestamp(&mut self, timestamp: Timestamp) -> &mut StdErrLog

Enables or disables the use of timestamps in log messages

pub fn color(&mut self, choice: ColorChoice) -> &mut StdErrLog

Enables or disables the use of color in log messages

pub fn module<T: Into<String>>(&mut self, module: T) -> &mut StdErrLog

specify a module to allow to log to stderr

pub fn modules<T: Into<String>, I: IntoIterator<Item = T>>(
&mut self,
modules: I
) -> &mut StdErrLog

specifiy modules to allow to log to stderr

pub fn init(&mut self) -> Result<(), SetLoggerError>

sets the the logger as active