Struct stderrlog::StdErrLog

pub struct StdErrLog { /* fields omitted */ }

Methods

impl StdErrLog

fn new() -> StdErrLog

fn verbosity(&mut self, verbosity: usize) -> &mut StdErrLog

Sets the verbosity level of messages that will be displayed

fn quiet(&mut self, quiet: bool) -> &mut StdErrLog

fn timestamp(&mut self, timestamp: Timestamp) -> &mut StdErrLog

Enables or disables the use of timestamps in log messages

fn module(&mut self, module: &str) -> &mut StdErrLog

fn modules<T: Into<String>, I: IntoIterator<Item = T>>(
&mut self,
modules: I
) -> &mut StdErrLog

fn init(&self) -> Result<(), SetLoggerError>